pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tnc_runoff"

description = "Fast runoff calcs for TNC via hspf2"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "BSD-3-Clause" }
authors = [{ name = "Austin Orr", email = "aorr@geosyntec.com" }]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Typing :: Typed",
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: BSD License",
]
dependencies = [
    "numpy",
    "pandas",
    "numba",
    "google-cloud-storage",
    "orjson",
    "pydantic",
    "pydantic-settings",
    "hsp2@https://github.com/Geosyntec/HSPsquared/archive/slim/859dc0d.zip",
]
dynamic = ["version"]

[tool.setuptools]
package-dir = { tnc = "tnc" }

[tool.setuptools.dynamic]
version = { attr = "tnc.__version__" }

[tool.ruff.lint]
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
